backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
email-validator==2.1.1
PyJWT==2.9.0           # najnowsza w repozytorium PyPI
passlib[bcrypt]==1.7.4
requests==2.32.2
openai==1.25.0
tweepy==4.14.0         # najwyższa dostępna wersja
stripe==10.0.0
weasyprint==61.0
python-multipart==0.0.9
typing-extensions==4.12.1
aiofiles==23.2.1
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
coverage==7.5.1
bandit==1.7.8
flake8==7.0.0
black==24.3.0
isort==5.13.2
mypy==1.10.0
